requirements.txt
dotenv
pypdf
fastapi
uvicorn
streamlit
requests
plotly
pandas
langgraph==1.0.4
langchain==1.1.3
langchain-core>=1.1.2,<2.0.0
langchain-text-splitters
langchain_community>=0.3.0
langchain_anthropic 
langchain-openai
langchain_experimental
langfuse
pymongo
pydantic>=2.0.0
psycopg2-binary
faiss-cpu